install stripe and add javascript for it on basket.html and a function createcheckoutsessionview to connect a stripe checkout

+++ Tickets/models.py
@@ -34,8 +34,7 @@ class Order(models.Model):
     order_date = models.DateTimeField(auto_now_add=True)
 
     # Add the total_price field
-    total_price = models.DecimalField(
-        max_digits=10, decimal_places=2, default=0)
+    total_price = models.IntegerField(default=0)  # cents
 
     def __str__(self):
         return f"{self.user.username} - {self.race.name}"
